# Change index to start from 1 by default in `predictions.json`