Closes https://github.com/ultralytics/ultralytics/issues/18136

The `category_id` in COCO JSON format starts from 1 in general, not just for LVIS. This is also the assumption used in the `convert_coco` function, and hence we subtract 1:
https://github.com/ultralytics/ultralytics/blob/d170f1ab6ad740369e5508788ab267a6639fc851/ultralytics/data/converter.py#L302

<!--
Thank you 🙏 for your contribution to [Ultralytics](https://www.ultralytics.com/) 🚀! Your effort in enhancing our repositories is greatly appreciated. To streamline the process and assist us in integrating your Pull Request (PR) effectively, please follow these steps:

1. Check for Existing Contributions: Before submitting, kindly explore existing PRs to ensure your contribution is unique and complementary.
2. Link Related Issues: If your PR addresses an open issue, please link it in your submission. This helps us better understand the context and impact of your contribution.
3. Elaborate Your Changes: Clearly articulate the purpose of your PR. Whether it's a bug fix or a new feature, a detailed description aids in a smoother integration process.
4. Ultralytics Contributor License Agreement (CLA): To uphold the quality and integrity of our project, we require all contributors to sign the CLA. Please confirm your agreement by commenting below:

    I have read the CLA Document and I sign the CLA

For more detailed guidance and best practices on contributing, refer to our ✅ [Contributing Guide](https://docs.ultralytics.com/help/contributing/). Your adherence to these guidelines ensures a faster and more effective review process.
--->


## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary  
Fixes a class mapping discrepancy to ensure consistent behavior across datasets like COCO and LVIS.

### 📊 Key Changes  
- Adjusted `class_map` to correctly map class indices for non-COCO datasets, starting from `1` instead of `0`.  
- Simplified category ID computation for LVIS by removing unnecessary conditional logic.  

### 🎯 Purpose & Impact  
- 🛠 **Purpose:** Resolves an indexing issue to align the `class_map` behavior with standard dataset conventions.  
- 🌍 **Impact:** Ensures accurate dataset processing for tasks using non-COCO datasets like LVIS, preventing incorrect metadata or evaluation errors.  
- 🤝 **User Benefit:** Improved reliability and consistency in YOLO validation workflows, especially for LVIS users.