@UltralyticsAssistant 2024-12-09
👋 Hello @Y-T-G, thank you for submitting a fantastic `ultralytics/ultralytics` 🚀 PR! To help us review and merge your work more efficiently, please ensure the following checklist is complete:

- ✅ **Define a Purpose**: Your detailed PR description is much appreciated! Thank you for linking [the relevant issue](https://github.com/ultralytics/ultralytics/issues/18136) and explaining the context. Excellent work providing clarity on the reasoning behind this change!  
- ✅ **Synchronize with Source**: Please confirm your PR is synchronized with the latest `main` branch of `ultralytics/ultralytics`. If it isn't already up to date, you can sync it by clicking the 'Update branch' button in your PR or by running `git pull` and `git merge main` locally.  
- ✅ **Ensure CI Checks Pass**: Double-check that all [Continuous Integration (CI)](https://docs.ultralytics.com/help/CI/) checks are passing for your PR. If any are failing, kindly address the issues.  
- ✅ **Update Documentation**: If necessary, please ensure that the [documentation](https://docs.ultralytics.com) has been updated to reflect your changes for affected datasets like COCO or LVIS.  
- ✅ **Add Tests**: If applicable, include or enhance the tests to verify the behavior of your changes, and ensure all tests are passing locally.  
- ✅ **Sign the CLA**: If this is your first time contributing to Ultralytics, kindly sign our [Contributor License Agreement](https://docs.ultralytics.com/help/CLA/) by commenting "I have read the CLA Document and I sign the CLA" below.  

🎯 **Precision and Impact:** Your changes streamline the handling of `class_map` for non-COCO datasets and improve alignment with dataset conventions. This is a great contribution for improving consistency in YOLO workflows! Please verify your modifications are as minimalistic as possible to resolve the issue without introducing unrelated changes, as per our [Contributing Guide](https://docs.ultralytics.com/help/contributing). 

If you encounter or identify any edge cases, feel free to share a **minimum reproducible example (MRE)** to help with validation and testing. This helps ensure that all aspects of the fix are covered robustly.  

An Ultralytics engineer will review your PR shortly to assist further. Thank you again for your contribution to Ultralytics! We can’t wait to see this improvement in action. 🚀✨