@glenn-jocher 2024-12-09
@Y-T-G thanks for the PR! Could you test val with `main` and this PR on other datasets like VOC that have zero-indexed classes? I'm not sure if this might break something there as I think all of the current YAMLs start from zero, .e.g:

https://github.com/ultralytics/ultralytics/blob/d170f1ab6ad740369e5508788ab267a6639fc851/ultralytics/cfg/datasets/VOC.yaml#L22-L44