@Y-T-G 2024-12-09
@glenn-jocher Sure. I think this only applies to COCO and LVIS because those are the only ones that trigger `eval_json` where we use the COCOAPI and the generated `predictions.json`. And also for DOTA to generate the predictions for submission to DOTA evaluation website. Otherwise, we do not make use of the `predictions.json` ourselves. It just get saved for the users when they use `save_json=True`.