@timecollision 2024-12-10
I change the code to 
cls = coco80[ann["category_id"]] if cls91to80 else ann["category_id"]  # class
but the "category_id" in predictions.json is still 0