@Y-T-G 2024-12-10
@timecollision That's for conversion from CCO to YOLO. If you want to test this PR, you can use:

```bash
git clone https://github.com/ultralytics/ultralytics
cd ultralytics
git checkout coco_pred_json
pip install .
```

Then run `model.val(save_json=True)`